refactor tests to test all csi driver api calls
